[Benja, Dani C] 178 Remove cypress folder, rename hoc folder, added persistence and folder structre entry in contributing
