+++ CONTRIBUTING/CONTRIBUTING.md
@@ -3,7 +3,7 @@
 - [General](/CONTRIBUTING/commands.md)
 - [Sobre nuestros endpoints](/CONTRIBUTING/endpoints.md)
 - [Sobre nuestros test](/CONTRIBUTING/test.md)
-- [(WIP) Capa de persistencia](/CONTRIBUTING/persistencia.md)
-- [(WIP) Arquitectura](/CONTRIBUTING/test.md)
+- [Sobre la capa de persistencia](/CONTRIBUTING/persistence.md)
+- [Sobre la estructura de archivos](/CONTRIBUTING/folder-structure.md)
 - [(WIP) Estilos](/CONTRIBUTING/test.md)
 - [(WIP) Server Side Rendering y Client Side Rendering](/CONTRIBUTING/test.md)
