+++ CONTRIBUTING/folder-structure.md
@@ -1,0 +1,91 @@
+# Folder structure
+
+Las carpetas más importantes son:
+
+## .circleci
+
+En esta carpeta se encuentran los archivos del código del pipeline de CircleCI.
+
+## .vscode
+
+Normalmente esta carpeta se ignora, pero dado que todo el equipo utiliza VS code para trabajar el proyecto, hemos decidido agregarlo al repositorio y subir algunos scripts para hacer debug.
+
+## CONTRIBUTING
+
+Carpeta con la documentación para desarrolladores.
+
+## migrations
+
+Migraciones de TypeORM.
+
+## pages
+
+Next JS creará una ruta por cada archivo que se encuentre en este directorio. Ver la documentación de [NextJS](https://nextjs.org/docs/routing/introduction)
+
+Las rutas de los endpoints se encuentran en `pages/api`
+
+Todas nuestras rutas de cara al usuario están en español porque se creo este sistema para hispanohablantes. Sin embargo, el código por debajo está todo en inglés.
+
+## public
+
+Se encuentran todos los archivos públicos tales como imágenes y fuentes.
+
+## src
+
+La carpeta `src` contiene todo el código funcional del sistema.
+
+### components
+
+Esta carpeta contiene los componentes de React que nosotros utilizamos en el proyecto.
+
+### dto
+
+Contiene los `types` para mover entidades entre servidor y cliente.
+
+### entity
+
+Especificación del mapeo entre las tablas de base de datos y el sistema.
+
+### high-order-function
+
+Se trata de funciones que miran aspectos del sistema. En este momento sólo tenemos un hof para validar la sesión de usuario, tanto a nivel de endpoints como a nivel de página.
+
+### model
+
+Aquí se encuentran los modelos de negocio del sistema.
+
+### services
+
+Reglas de negocios del sistema.
+
+### steps
+
+Estos son los componentes root que se renderizan cuando se carga una ruta en particular de pages.
+
+### styles
+
+Definición de constantes para SASS.
+
+### theme
+
+Definición de tema para ChakraUI.
+
+## test
+
+En tests se encuentran todos los archivos que se relacionan a la automatización de pruebas.
+
+### \_\_endpoints
+
+Contiene archivos http que son archivos estándar para poder ejecutar rápidamente peticiones http y ejecutar los endpoints de forma más sencilla.
+
+### **mocks**
+
+Mock que ocupa jest.
+
+### pages
+
+En general aquí es donde escribirás la mayor parte de tus tests, ya que todos nuestros tests prueban una ruta en particular y todas las funcionalidades que ésta tiene. Por cascada, se prueba todo el código asociado a esa ruta.
+
+### src
+
+Hay código que no necesariamente se puede ejecutar a través de pages como por ejemplo los hof que deben ser probados aparte.
