--- cypress/page-objects/FaceScaleScreen.js
@@ -1,7 +1,0 @@
-class FaceScaleScreen {
-  clickPainLevel(level) {
-    cy.get(`img[alt="${level}"]`).click()
-  }
-}
-
-export default FaceScaleScreen
