--- cypress/page-objects/SuccessfulSymptomsRegistry.js
@@ -1,11 +1,0 @@
-class SuccessfulSymptomsRegistry {
-  visit() {
-    cy.visit('/successful-symptoms-registry')
-  }
-
-  clickGoBackHome() {
-    cy.contains('Ok, volver al inicio').click()
-  }
-}
-
-export default SuccessfulSymptomsRegistry
