--- cypress/page-objects/WelcomeSayu.js
@@ -1,11 +1,0 @@
-class WelcomeSayu {
-  visit() {
-    cy.visit('/')
-  }
-
-  clickRegisterSymptomsButton() {
-    cy.contains('Registrar síntomas').click()
-  }
-}
-
-export default WelcomeSayu
