+++ src/steps/SymptomsRegistry/SymptomsRegistry.tsx
@@ -13,7 +13,7 @@ import ButtonLink from 'src/components/ButtonLink'
 import { ErrorCodes } from 'src/components/Error'
 import { SuccessCodes } from 'src/components/Success'
 
-import withSession from 'src/hoc/WithSession'
+import withSession from 'src/high-order-function/WithSession'
 import { Role } from 'src/model/Role'
 
 import styles from './SymptomsRegistry.module.scss'
