+++ src/steps/WelcomeSayu/WelcomeSayu.tsx
@@ -4,7 +4,7 @@ import { signOut, useSession } from 'next-auth/client'
 import { Flex, Image, Text, Stack } from '@chakra-ui/core'
 
 import ButtonLink from 'src/components/ButtonLink'
-import withSession from 'src/hoc/WithSession'
+import withSession from 'src/high-order-function/WithSession'
 import { Role } from 'src/model/Role'
 import styles from './WelcomeSayu.module.scss'
